Don't force PreferencesFactory initialization
It instantiates a few parts of infrastructure early,
before parts of Setup.php have been executed, making
service initialization risky. It also forces initialization
of a few dependencies, even if PF itself will never be
required during this request.

If there are two extensions that want to redefine this
service, there is a bigger problem and trying to solve
it by stomping on the competition is not going to help.

Bug: T190425
Depends-On: I749f66d13a8c8a8ae4a83661b83c56f0db74a187
Change-Id: I314b6f9117ec8f73145b569e4ad6e5cc2ea97051