Don't delete all global preferences when saving
This is to protect those prefs that may have been set by
extensions that are not present on the current wiki.

All global preferences are still deleted when resetting;
the Storage::delete() method now accepts an optional
array of preference names to work with, and if left out
all for the user will be deleted.

Bug: T186191
Change-Id: I9dab3da202f24a8308ffdd8e7cd024dade032f7b