Fix handling for CentralIdLookup::namesFromCentralIds()
When saving global preferences we have to get the list of the
user's preferences. In doing so we retrieve that user's
options, but at that point they've not been saved and so haven't
been converted to their final form. Hence the error. This is
avoided by switching to a copy of the user (with old option
values).

Also, an extra check is brought in from User::saveOptions()
that will also avoid the same problem, if arrays are passed in
by the API.

Bug: T190353
Change-Id: I130409ab0495d162983074985addcbc6d00f3693