Prevent "staggered" checkbox fields at some font sizes / zoom levels
Adding `clear: left;` to the existing `float: left;` rule will ensure
that the browser will never place global preference checkboxes side by
side, which causes the indentation/stairs/staggered appearance.

Nothing else on the preferences page is floated, so the layout of
other elements of the page is not affected.

Bug: T190355
Change-Id: I544dcc5fbf7b0765d02d6e450bfef3a0643c1356