OOUI'fy Special:GlobalPreferences
This will allow the page to work with the ?ooui=1 parameter
and convert to OOUI elements when OOUI mode is enabled.

Also allows the system to toggle the individual preference
widgets enabled/disabled based on the Global Preference
checkbox, and toggle the "select all" checkbox in each
section per the state of the other individual preferences'
checkbox in that section.

Bug: T186842
Depends-On: I9943b0aa1746fdfb60c7d4c88d6d4d7ac0589a2c
Change-Id: Icc55192bac32bb16ca6a8dd2cdf2659984b725b3