Improve "option is globally overridden" error message
The current error message is a dead end. It's not even really
clear what it refers to, because you see this when you try to
use the regular "options" API. You have no idea it refers to
the GlobalPreferences extension, except you already understand
what "globally overridden" refers to.

For a user of the API I think these are the 2 possible ways
forward.

Bug: T271434
Change-Id: Ia05a952860bf652c9805f0f72c7d0bf3caab71de