Handle loosly-boolean local exception values
The values of the local-exception preferences were assumed to be '1'
or null, but actually in some circumstances can also be '0' (which
should be assumed to be the same as null, i.e. no-local-exception set).

This change makes the various checks of these prefernces values more
lenient, and also in some places avoids the deprecated User methods.

Bug: T286271
Change-Id: I0f585db79a4c89f7707ce3d7ee8599d8e89dfa97