Convert to HookHandlers
This converts all hooks to implementations of their respective classes.
The MediaWikiServicesHook is moved into its own class because it can't
have other services injected. The LoadExtensionSchemaUpdatesHook is
the same, but it uses Config and so for now I've left it not having that
injected.

Other changes in Hooks.php are simplifying where possible, now that
various things are injected.

Change-Id: Ifcad7d4f7e05a0a9f5bf7b8ff49329981ad00099