tests: Explicitly disable main cache in one test suite
Most caches in core are enabled in PHPUnit, through a dedicated
in-memory HashBagOStuff that is cleared before each test. An exception
until now was WANObjectCache which, while normally wrapping the "main"
cache (which is enabled in PHPUnit), was explicitly wrapping "none"
instead.

I'm fixing this in I53781a8c06ebb2583f6ca8 in core, but noticed
these failures in GlobalPreferencesTest. These tests ideally would
work correctly under prod-like circumstances and excercise their
cache interaction accordingly. A fix might involve splitting tests
into separate test cases, or explicitly resetting services mid-way
a test (and thus reset your Storage service and its copy of the
cache service), or if there is a chance that this is a bug in prod
code then perhaps certain methods may be incorrectly calling on the
cache during a write request. For example, it may be incorrectly
assuming that a purge is instantaneous across all servers, DCs, and
on-going processes.

```
1) GlobalPreferences\Test\GlobalPreferencesTest::testStorage
   Failed asserting that actual size 1 matches expected size 2.
   GlobalPreferences/tests/phpunit/GlobalPreferencesTest.php:52
==Logs==
   [objectcache] fetchOrRegenerate(global:globalpreferences:prefs:1:5): volatile hit

2) GlobalPreferences\Test\GlobalPreferencesTest::testUserPreference
   Failed asserting that two strings are equal.
   -'ru'
   +'de'
   GlobalPreferences/tests/phpunit/GlobalPreferencesTest.php:93
==Logs==
   [objectcache] fetchOrRegenerate(global:globalpreferences:prefs:1:6): volatile hit
   [objectcache] fetchOrRegenerate(global:globalpreferences:prefs:1:6): volatile hit

3) GlobalPreferences\Test\GlobalPreferencesTest::testAutoGlobals
   Failed asserting that two strings are equal.
   -'ru'
   +'de'
   GlobalPreferences/tests/phpunit/GlobalPreferencesTest.php:122
== ogs==
   [objectcache] fetchOrRegenerate(global:globalpreferences:prefs:1:7): volatile hit
   [objectcache] fetchOrRegenerate(global:globalpreferences:prefs:1:7): volatile hit
```

Change-Id: Ie1def75208822bdf19bb2cfd7e6edf32c2000e6b