GlobalPreferencesFactory: Use native way to add extra help message
The `$context->msg()` method is not compatible with the `help-message`
param for HTMLForm when message parameters supplied.
Use proper params like `help-messages` for HTMLForm to handle the
message parsing instead.

This also can improve performance of the `options` API endpoint,
since it don't need the help messages to be parsed.

Bug: T337278
Change-Id: If2e6ed5c0cc616e53393e01932105a9b15fec76f