ApiGlobalPreferences: allow resetting of preferences
Storage::save() does the actual work of deleting rows. The previous
implementation used a class property to track which preferences to
reset, but in doing so unsets them from the array that gets passed to
Storage::save(), thus it never was actually deleted from the database.

Similarly, preferences that are reset via the 'reset' and 'resetkinds'
options should have their values set to null, and not unset from the
array passed to the storage layer.

Add new tests for ApiGlobalPreferences class.

Bug: T207448
Change-Id: I2f1affb170f4942ed0d2124b1ef8f6e3287b8b5c