Use new ApiOptionsBase base class
I split ApiOptionsBase off from ApiOptions so that core is able to
change things without affecting the extension.

There are some minor changes to how dependencies are injected and used.

The core change also allowed the hook call to be omitted for
action=globalpreferences, so in the hook handler, we don't need to
check for that anymore.

Bug: T323076
Depends-On: I53e5c409a650397fde03a8578b0182f0b97927a9
Change-Id: I686c0091768ed330683fff89201fae4f93b8a1c4