Implement the new fetchBatchForUserNames method
Support the global gender feature by implementing the new batch fetch
method.

Use the new method CentralIdLookup::lookupOwnedUserNames() in both
the batch and single cases.

In the single case, this avoids very bad worst-case performance in
isOwned() if the user's CentralAuth WAN cache record is missing.
centralIdFromName() was already doing a query, and that's apparently
unavoidable, so sharing that query with the attachment check is always a
win.

Bug: T386584
Depends-On: I99dd733547f46e635c2f363166ce111e76e06a5e
Change-Id: Id02489a597f96cd1cd6db08e16b3624542fdf1f7