Call parent::execute() after SpecialGlobalPreferences checks
In SpecialGlobalPreferences::execute(), the parent::execute() call was
moved from the end to the start in Icc55192bac32bb16ca6a8dd2cdf2659984b
because it was desired to access $this->oouiEnabled but it was
initialised by the parent. However, this change defeated the purpose of
half the method. It is not useful to check whether the user is logged
in and globalised after displaying the whole form.

The effect was a RuntimeException for non-global users which is now
causing CI to fail against master.

So, restore the placement of the parent::execute() call. The oouiEnabled
property no longer exists, so that's not a problem anymore.

Also, memoize GlobalPreferencesFactory::getUserID(). To avoid this
exception, it's important that the same user ID is used every time this
method is called.

Add integration test.

Bug: T314741
Change-Id: I24b17d9484a67511fae8ecaa3bcba925c8fec318